Add check for confirmation count in dcrdata
This adds a confirmation count check for the funding transaction by
querying dcrdata for the tx confirmation count every minute after a
successful channel opening.

This is lame, but sufficient since a check every minute is pretty light
on resource usage and it should stop after very few checks anyway (~
checks or so).